Add basic Android module
This commit adds the base for the Android part, written in Kotlin.

The module exposes the function `openInApp` which currently does
nothing: it is there merely a placeholder to allow the TS part to
connect to the module.

Note that with the addition of the types for `react-native`, some
definitions where conflicting with `@types/node`, which in turn was
removed.