Add pre and post package scripts
In order to create a usable package, some steps need to be made:

1. Compile Typescript to Javascript
2. Separate type definitions into dedicated directory
3. Move Android and iOS folder to project root for `react-native link`
4. Remove unnecessary entries from package.json

Steps 1 is accomplished with the `build` script. Step 2 and 3 are
carried out by `scripts/prepack.ts`. The two scripts are chained
together, along with `test` in the `prepackade` script.

Finally, step 4 and necessary cleanup is done with
`scripts/postpack.ts` which is triggered by the `postpackage` script.

Finally, the husky pre-commit hook has been temporarily removed as it
seems that `precise-commits` cannot handle formatting partial json
files.