Add iOS module
This commit adds the iOS part, written in Swift.

The module exposes a `openInApp` function that mimics the behaviour of
the existing Android one: it attempts to open a `SFSafariViewController`
view (rough equivalent of Chrome Custom Tabs) if the OS supports it,
otherwise it falls back to opening the link the the default browser.

While Chrome Custom Tabs handles the fallback internally,
`SFSafariViewController` does not, so it's fallback logic is manually
defined.

Also, the pre-pack script and other files have been updated to properly
deal with the added iOS project.