Add fallback dependency versions for Android
Android Support Library dependencies must all be of the same version
across the entire application. For this reason, in the `build.gradle`
file the dependencies would get version numbers from the root
application project. However, there was no fallback for when the
versions are not defined in the root project.

This commit adds a function that will return the dependency version
number as defined in the root project, else falls back to a default
value.

This brings better chances for an app to compile even if the generated
application Gradle configuration by React Native is edited.

As a side effect, Working on this library from Android Studio will
provide code completion because now Gradle has dependency version
default values to work with, so it will know what to download.