Add default settings initialization function
The settings now has a `defaultSettings` object with can be initialized
with the new `initialize`.

This utility can be quite useful for scenarios where `openInApp` is used
in several places and the settings are always the same: inizializing
would mean that `openInApp` would need only the mandatory url and
nothing more.