Fix settings sanitization immutability issue
Since Object assignment in Javascript is by reference and not by value,
the `sanitiziedSettings` initialization in both sanitiziation functions
would actually modify the `defaultSettings` which should be immutable.

This commit fixes this issue by initializing `sanitizedSettings` with
the object spread operator so to not create a reference but create a new
object instead.