Upgrade peer dependencies
ESLint is upgraded to the most recent v5 version, not v6. This
limitation is due to other dependencies, like
`@react-native-community/eslint-config`, that still rely on the previous
version.

TypeScript is upgraded to v3.7. Some portions of the code is updeated
to use the new optional chaining and the null coalescing operator.

Other dependencies are also updated to the their most recent versions.